+++ sdk/src/azure/iot/az_iot_provisioning_client.c
@@ -37,11 +37,7 @@ AZ_INLINE az_span _az_iot_provisioning_get_str_dps_registrations()
   return az_span_slice(_az_iot_provisioning_get_dps_registrations_res(), 0, 19);
 }
 
-AZ_NODISCARD az_result _iot_provisioning_client_parse_operation_status(
-    az_span response_operation_status,
-    az_iot_provisioning_client_operation_status* out_operation_status);
-
-    AZ_NODISCARD az_iot_provisioning_client_options az_iot_provisioning_client_options_default()
+AZ_NODISCARD az_iot_provisioning_client_options az_iot_provisioning_client_options_default()
 {
   return (az_iot_provisioning_client_options){ .user_agent = AZ_SPAN_EMPTY };
 }
@@ -343,7 +339,7 @@ AZ_INLINE az_result _az_iot_provisioning_client_payload_registration_state_parse
   return AZ_OK;
 }
 
-AZ_NODISCARD az_result _iot_provisioning_client_parse_operation_status(
+AZ_NODISCARD static az_result _az_iot_provisioning_client_parse_operation_status(
     az_span response_operation_status,
     az_iot_provisioning_client_operation_status* out_operation_status)
 {
@@ -418,7 +414,7 @@ AZ_INLINE az_result az_iot_provisioning_client_parse_payload(
       {
         return AZ_ERROR_ITEM_NOT_FOUND;
       }
-      _az_RETURN_IF_FAILED(_iot_provisioning_client_parse_operation_status(
+      _az_RETURN_IF_FAILED(_az_iot_provisioning_client_parse_operation_status(
           jr.token.slice,
           &out_response->operation_status));